WhenThereAreExceptionsInSqsCalling
